// src/models/viewmodels/ControlIngreso/ControlIngresoEntity.ts
export interface ControlIngresoEntity {
    idCompra: number;
    fechaCompra: string;
    deposito: number;
    depositoDescripcion: string;
    nroFactura: string;
    idOrden: number;
    nroProveedor: number;
    proveedor: string;
    proveedorId: number;
    verificado: number;
    responsableUbicacion: number;
    estado: string;
}